[jit] move private implementation out of `jit/__init__.py`
We pack a lot of logic into `jit/__init__.py`, making it unclear to
developers and users which parts of our API are public or private. This
is one in a series of PRs intended to pull implementation out into
separate files, and leave `__init__.py` as a place to register the
public API.

I'm going to try to make us conform to the relevant rules in [PEP 8](https://www.python.org/dev/peps/pep-0008/#public-and-internal-interfaces).

[ghstack-poisoned]
